+++ messaging/task.py
@@ -11,8 +11,8 @@ from rest_framework import status
 from rest_framework.generics import get_object_or_404
 
 from messaging.models import Message, MessageStatus, Channel, MessageDirection
-from messaging.serializers import MessageData, MessageSerializer
-from messaging.views import send_bulk_message
+from messaging.serializers import NotificationData, MessageSerializer
+from utils.notification import send_bulk_notification
 
 
 class CommCareHQAPIException(Exception):
@@ -70,21 +70,20 @@ def delete_old_messages():
     """
       Deletes messages that are older than 7 days.
     """
-    print("Running delete_old_messages")
     cutoff_date = now() - timedelta(days=7)
     deleted_count, _ = Message.objects.filter(received__lte=cutoff_date).delete()
 
 
 @shared_task(name="resend_notifications_for_undelivered_messages")
 def resend_notification_of_undelivered_messages():
-    undelivered_msgs = (Message.objects.filter(status=MessageStatus.SENT_TO_SERVICE,
-                                               direction=MessageDirection.SERVER).select_related('channel',
+    undelivered_msgs = (Message.objects.filter(recevied__isnull=True,
+                                               direction=MessageDirection.MOBILE).select_related('channel',
                                                                                                  'channel__connect_user'))
     for msg in undelivered_msgs:
         channel = msg.channel
         serialized_msg = MessageSerializer(msg)
-        message_to_send = MessageData(
+        message_to_send = NotificationData(
             usernames=[channel.connect_user.username],
             data=serialized_msg.data
         )
-        send_bulk_message(message_to_send)
+        send_bulk_notification(message_to_send)
